Keep offline artifact resolution internal
